create layout for additional Help info
