+++ src/App.tsx
@@ -17,14 +17,13 @@ import BossSummary from './uicomponents/BossSummary.tsx';
 import Navbar from './uicomponents/Navbar.tsx';
 import RaidControls from './uicomponents/RaidControls.tsx';
 import LinkButton from './uicomponents/LinkButton.tsx';
+import StratHeader from './uicomponents/StratHeader.tsx';
+import StratFooter from './uicomponents/StratFooter.tsx';
 
 import { Generations, Pokemon, Field} from './calc/index.ts';
 import { MoveName } from './calc/data/interface.ts';
 import { Raider, RaidBattleInfo, RaidState } from './raidcalc/interface.ts';
-import StratHeader from './uicomponents/StratHeader.tsx';
-import StratFooter from './uicomponents/StratFooter.tsx';
 
-// import {BOSS_SETDEX_SV} from './data/sets/raid_bosses.ts'
 
 function App() {
   const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)');
